requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
easyocr==1.7.0
filelock==3.12.2
fonttools==4.42.1
gunicorn==21.2.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.7
pyclipper==1.3.0.post4
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2023.8.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
